refactor: move fetchUtkast from utkastFetch.ts to utkastUtils.ts
Consolidates fetch logic into the existing utils module, reducing the
number of files in the utkast feature slice.
